services/relay/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
pyserial==3.4
requests==2.23.0
urllib3==1.25.8
pytz==2019.1
pycrypto==2.6.1
miniupnpc==2.0.2